Cleanup & review from Microwindows git v0.94 9/28/17 to github version on 11/23/18

+++ src/mwin/winlib/scrlbar.c
@@ -1027,7 +1027,7 @@ ScrollbarControlProc (HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam)	/* 
 #if 0	/* jmt: must handle WM_MOVE */
 		pData->rc=hwnd->winrect;   /* screen coordinates position*/
 #endif
-            	hwnd->userdata = (DWORD)pData;
+            	hwnd->userdata = pData;
 
 		if (dwStyle == SBS_VERT)
 		{
